添加支付，二维码sdk，添加支付确认选择界面
添加支付，二维码sdk，添加支付确认选择界面

+++ HealthyCertificate/ProjectClass/Model/WaitToCheckModel/BRContract.h
@@ -74,7 +74,7 @@
 /**
  * 预约人数
  */
-@property (nonatomic, assign) int regCheckNum;
+@property (nonatomic, assign) NSInteger regCheckNum;
 /**
  * 已检人数
  */
